python/requirements.txt
Django==2.2.24
uwsgi==2.0.18
mysqlclient==1.4.6
django-bootstrap4
django-widget-tweaks
django-mysql==3.8.1
pyjq
requests
six
stripe
qrcode
pillow
python-dateutil==2.8.1
beautifulsoup4==4.11.1
